+++ django/applications/catmaid/static/js/layers/pixi-tile-layer.js
@@ -113,6 +113,18 @@
         scaledStackPosition.z,
         scaledStackPosition.s);
 
+    if (this.hideIfNearestSliceBroken) {
+      // Re-project the stack z without avoiding broken sections to determine
+      // if the nearest section is broken.
+      var linearStackZ = this.stack.projectToLinearStackZ(
+          this.stackViewer.projectCoordinates().z);
+      if (this.stack.isSliceBroken(linearStackZ)) {
+        this.batchContainer.visible = false;
+      } else {
+        this.setOpacity(this.opacity);
+      }
+    }
+
     var effectiveTileWidth = this.tileSource.tileWidth * tileInfo.mag;
     var effectiveTileHeight = this.tileSource.tileHeight * tileInfo.mag;
 
@@ -274,28 +286,6 @@
     this._oldZ    = this._swapZ;
 
     this._renderIfReady();
-  };
-
-  /**
-   * Returns a set of set settings for this layer. This will only contain
-   * anything if the tile layer's tile source provides additional settings.
-   */
-  PixiTileLayer.prototype.getLayerSettings = function () {
-   if (this.tileSource && CATMAID.tools.isFn(this.tileSource.getSettings)) {
-     return this.tileSource.getSettings();
-   } else {
-     return [];
-   }
-  };
-
-  /**
-   * Set a layer setting for this layer. The value will only have any effect if
-   * the layer's tile source accepts setting changes.
-   */
-  PixiTileLayer.prototype.setLayerSetting = function(name, value) {
-   if (this.tileSource && CATMAID.tools.isFn(this.tileSource.setSetting)) {
-     return this.tileSource.setSetting(name, value);
-   }
   };
 
   CATMAID.PixiTileLayer = PixiTileLayer;